requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-cloud-storage>=2.14.0
python-dotenv>=1.0.0
flask>=3.0.0